Refactor PlayerExternal namespaces
