+++ Packages/org.centurioncc.system/Runtime/Player/External/HitDisplay/ExternalHitDisplay.asset
@@ -21,7 +21,7 @@ MonoBehaviour:
   compiledVersion: 2
   behaviourSyncMode: 1
   hasInteractEvent: 0
-  scriptID: -4802658992678766763
+  scriptID: 2259173546374559232
   serializationData:
     SerializedFormat: 2
     SerializedBytes:
