+++ Packages/org.centurioncc.system/Runtime/Player/External/HitDisplay/ExternalHitDisplayManager.cs
@@ -3,7 +3,7 @@ using UdonSharp;
 using UnityEngine;
 using VRC.SDKBase;
 
-namespace CenturionCC.System.Player.PlayerExternal
+namespace CenturionCC.System.Player.External.HitDisplay
 {
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
     public class ExternalHitDisplayManager : PlayerManagerCallbackBase
