+++ Packages/org.centurioncc.system/Runtime/Player/External/PlayerTag.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6c5c1f5c89e815649a9eb5af3cb8818e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
