+++ Packages/org.centurioncc.system/Runtime/Player/External/PlayerTag/ExternalPlayerTag.asset
@@ -21,7 +21,7 @@ MonoBehaviour:
   compiledVersion: 2
   behaviourSyncMode: 1
   hasInteractEvent: 0
-  scriptID: 3484353473929384646
+  scriptID: 4709086280642371146
   serializationData:
     SerializedFormat: 2
     SerializedBytes: 
@@ -722,7 +722,8 @@ MonoBehaviour:
       Data: 41|System.RuntimeType, mscorlib
     - Name: 
       Entry: 1
-      Data: CenturionCC.System.Player.PlayerExternal.ExternalPlayerTagManager, CenturionCC.System
+      Data: CenturionCC.System.Player.External.PlayerTag.ExternalPlayerTagManager,
+        CenturionCC.System
     - Name: 
       Entry: 8
       Data:
