+++ Packages/org.centurioncc.system/Runtime/Player/External/PlayerTag/ExternalPlayerTag.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 using UnityEngine.UI;
 using VRC.SDKBase;
 
-namespace CenturionCC.System.Player.PlayerExternal
+namespace CenturionCC.System.Player.External.PlayerTag
 {
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
     public class ExternalPlayerTag : ExternalPlayerTagBase
